Moved the all files access prompt from the LoadingActivity to the AccessActivity
